requirements.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.1.0
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
diff-match-patch==20121119
Django==1.11.15
django-crispy-forms==1.7.2
django-crontab==0.7.1
django-filter==1.1.0
django-formtools==2.1
django-import-export==1.0.1
django-redis==4.9.0
django-reversion==2.0.13
django-seed==0.1.9
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
drf-extensions==0.3.1
elasticsearch==2.4.1
et-xmlfile==1.0.1
Faker==0.9.0
future==0.16.0
httplib2==0.9.2
idna==2.6
itsdangerous==0.24
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
kombu==4.2.1
MarkupSafe==1.0
odfpy==1.3.6
openpyxl==2.5.4
Pillow==5.1.0
pycryptodomex==3.6.4
PyJWT==1.6.4
PyMySQL==0.8.1
python-alipay-sdk==1.7.1
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
redis==2.10.6
requests==2.20.0
six==1.11.0
tablib==0.12.1
text-unidecode==1.2
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.23
vine==1.1.4
xadmin==0.6.1
xlrd==1.1.0
xlwt==1.3.0
django-bootstrap3==11.0.0
django-cors-headers==2.4.0
gunicorn==19.6.0
pycryptodome==3.7.3
